requirements.txt
Pyrogram==2.0.30
TgCrypto==1.2.3
aiohttp
python-dotenv
beautifulsoup4==4.11.1
requests==2.28.1
aiofiles
translators